pyproject.toml
[project]
name = "language-companion"
version = "0.1.0"
description = "AI-powered language learning companion built with LangChain and Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit",
  "langchain",
  "langchain-core>=0.2",    
  "langgraph",
  "langchain-community",
  "langchain-openai",
  "tiktoken",
  "pydantic>=2",
  "python-dotenv",           
  "pymupdf>=1.26.5",         
  "python-docx>=1.2.0",       
]

[tool.uv]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","B","I","UP"]
ignore = ["E203","E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[dependency-groups]
dev = ["ruff", "black"]
